Level-1
1) renamed file back to Duke.
2) Added echo.
